Large refactor of Input handling
The input handler was doing too much. Editor actions were split into
their own files.

The input handler was refactored to raise events of things that occur
based on UnityEngine.Event and the nodes in the canvas. e.g. NodeClicked,
CanvasClicked.

The editor is subscribed to these events and executes the appropriate
action. e.g. Drag, Area Select, etc.

Selection of nodes and references was moved to its own class. This
handles setting the active objects and raising events when selections
are changed.

This greatly decoupled the Bonsai Designer code but there are still a few
things left. The goal is to decouple the need to have BonsaiWindow
references in classes such as Coord, BonsaiEditor, and BonsaiInput.