Rework Action handling
All editing functions are handled in the Editor.
The editor is the controller.

The BonsaiInput is responsible for generating event that are key to the
editor. e.g Node clicked or unclicked.

Moved drawing logic to BonsaiViewer (out of Editor)

Implemented Extended Selection (CTRL+Click to toggle selection).

Node linking changed. No longer done in inspector. SHIFT + Click to link
nodes.

Fixed guard inspector to update when key values change in one guard. The
link guards array is now hidden from the inspector to prevent invalid
state. It now shows the total number of guards linked instead. It is
kind of pointless to expose the array. Cant really do much besides set
it to null or remove entries.