Replace foreach with for in critical sections
Drawing nodes uses for loop.

Querying nodes under mouse uses for loop.

The canvs nodes are exposed in the public API as a IReadOnlyList. This
allows the Input system to take a random access read only list to
traverse in reverse without an IEnumerable. Reverse traversal is
required so the "top (last)" rendered node recieves input first.