Add blackboard if missing when opening asset
This is a safeguard for trees that have null blackboards when created via the CreateAssetMenu.

I could not find a direct method that would only execute if the Asset is created from the Create Menu. I did not want to pollute initialization code in the BehaviourTree class to automatically add the blackboard. It would get messy and require UNITY_EDITOR only code block.

To work around this, blackboards can be added directly to the tree from the context menu from the inspector. They are also added automatically when opened in Bonsai Designer.