Remove SaveAs action
SaveAs was incomplete and only did a asset copy.

Doing a proper saveAs is not as trivial if the original source should be
left unmodified and the copy with the latest canvas changes due to how
the AssetDatabase saves.

The BonsaiCanvas would need to keep clones of the node and tree that are
not tied to the AssetDatabase and saving would use the clones to apply
changes to the asset database.

SaveAs should be re-added once a correct implementaiton is made.

Assets can still be duplicated in the Unity editor.