Fix multi-select edit and title content
When editing multiple nodes at once, all selected node GUIs are updated.

Window title content updates on GUI.
This simplified logic to handle showing the update to date tree name in
the title.

This was also done because OnProjectChange message for
EditorWindow was sometimes called before the Tree Asset had the correct
name after the asset was renamed in the Project.