Rework Core API
This is an attempt to reduce state and simplify implementation of core Tree and Nodes.

The public API only allows nodes to be added in pre-order.

Simplified API to add and remove node children.

Split connected and unused nodes in tree asset storage. Unused nodes (which are not connected to main tree root) are kept in a separate list.

Added additoinal tests for Tree Cloning.