Change blackboard observer API
To prevent GC, blackboard adds a simple, but specialized interface for
observering blackboard changes.

Blackboard.IObserver is now used to get around the GC issue. This
allows nodes to implement themselves as an obeserver.

The main cause for GC was the garbage created from closures when using
Action<T> as the observer function.